feat: steel thread for standard queues
[#186768674](https://www.pivotaltracker.com/story/show/186768674)

An alternative implementation of the work done in:
https://github.com/cloudfoundry/csb-brokerpak-aws/pull/1503

We as a team decided it was worth experimenting with this
alternative approach for the steel thread for FIFO queues story
[#186768689](https://www.pivotaltracker.com/story/show/186768689)
but since its description clearly states "Build on #186768674"
I felt I needed to extend the experiment to that story as well.

This experiment uses terraform-aws-sqs as a submodule because
it suits better our existing workflow, however it shouldn't be
necessary at all since modules can simply point at gh repos:
https://developer.hashicorp.com/terraform/language/modules/sources#github

Should we decide that we want to remove the submodule we would
need to modidy our brokerpak packaging scripts to ensure that
we embed a local copy of the module (which is not hard to do).