feat: lock modules and providers versions (as much as possible)
[#186768674](https://www.pivotaltracker.com/story/show/186768674)

This commit should fix some existing problems we have seen in
our tests. We saw that terraform tests were using a different
version of the AWS Provider than, for example, tile tests.

By simply including this .terraform.lock.hcl file, which gets
generated automatically the first time we run `terraform init`,
we render such kind of version mismatches impossible.

When we (or csb) invokes `terraform init`:
1. If it doesn't find a .terraform.lock.hcl
  1. Looks for a version satisfying constraints in versions.tf
  1. I BELIEVE it optimises for cache, preferring local copies
  1. If can't find a local copy it downloads the latest one
     (which, by definition, can vary from day to day)
1. If it finds a .terraform.lock.hcl
  1. Looks for the version specified in .terraform.lock.hcl
  1. It DEFINITELY optimises for cache, preferring local copies
  1. If can't find a local copy it downloads that EXACT version

Downloads can be disabled to ensure a self-contained brokerpack

I chose to specify v5.33.0 for the provider because that's the
version we embedded in latest csb-aws v1.8.0.
https://network.pivotal.io/products/cloud-service-broker-aws/#/releases/1458843

Bumping versions in .terraform.lock.hck file to use a newer
version of a provider is as simple as `terraform init -upgrade`

To ensure that our releases works as expected and they bundle
the right version of every provider we would need to also
prevent any on-the-fly downloads altogether.
We might be doing it already in cloud-service-broker core.
By doing this we also ensure that terraform and tile-tests
will fail if terraform binary can't find a local copy of
a required provider.

Disclaimer:
Unless we remove the submodule and specify the module source as
a GitHub url as described in the official documentation
https://developer.hashicorp.com/terraform/language/modules/sources#github
the modules.json will just be pointing to a relative path which
is not as good for traceability. No explicit semver information